split up big default component
