+++ frontend/src/styles/HetComponents/HetTags.tsx
@@ -1,0 +1,19 @@
+interface HetTagsProps {
+  tags: string[]
+}
+
+export const HetTags: React.FC<HetTagsProps> = ({ tags }) => {
+  return (
+    <div className="md:flex md:flex-wrap mt-2 hidden">
+      {tags.map((name, index) => (
+        <span
+          aria-label={name}
+          key={index}
+          className="category-span text-[10px] uppercase text-altBlack font-sansTitle font-bold bg-ashgray30 rounded-sm py-1 px-2 mr-2 mt-1"
+        >
+          {name}
+        </span>
+      ))}
+    </div>
+  );
+}
